+++ chia-blockchain-gui/src/components/core/components/Amount/Amount.tsx
@@ -9,7 +9,7 @@ import {
 } from '@material-ui/core';
 import { useWatch, useFormContext } from 'react-hook-form';
 import TextField, { TextFieldProps } from '../TextField';
-import { chia_to_mojo } from '../../../../util/chia';
+import { chia_to_pupps } from '../../../../util/chia';
 import useCurrencyCode from '../../../../hooks/useCurrencyCode';
 import FormatLargeNumber from '../FormatLargeNumber';
 import Flex from '../Flex';
@@ -40,7 +40,7 @@ function NumberFormatCustom(props: NumberFormatCustomProps) {
 }
 
 export type AmountProps = TextFieldProps & {
-  children?: (props: { mojo: number; value: string | undefined }) => ReactNode;
+  children?: (props: { pupps: number; value: string | undefined }) => ReactNode;
   name?: string;
 };
 
@@ -54,7 +54,7 @@ export default function Amount(props: AmountProps) {
     name,
   });
 
-  const mojo = chia_to_mojo(value);
+  const pupps = chia_to_pupps(value);
 
   return (
     <FormControl variant={variant} fullWidth={fullWidth}>
@@ -71,18 +71,18 @@ export default function Amount(props: AmountProps) {
         }}
         {...rest}
       />
-      {!!mojo && (
+      {!!pupps && (
         <FormHelperText>
           <Flex alignItems="center" gap={2}>
             <Flex flexGrow={1} gap={1}>
-              <FormatLargeNumber value={mojo} />
+              <FormatLargeNumber value={pupps} />
               <Box>
-                <Plural value={mojo} one="mojo" other="mojos" />
+                <Plural value={pupps} one="pupps" other="puppss" />
               </Box>
             </Flex>
             {children &&
               children({
-                mojo,
+                pupps,
                 value,
               })}
           </Flex>