+++ chia-blockchain-gui/src/components/plotNFT/select/PlotNFTSelectPool.tsx
@@ -11,7 +11,7 @@ import PlotNFTSelectBase from './PlotNFTSelectBase';
 import normalizeUrl from '../../../util/normalizeUrl';
 import getPoolInfo from '../../../util/getPoolInfo';
 import InitialTargetState from '../../../types/InitialTargetState';
-import { chia_to_mojo } from '../../../util/chia';
+import { chia_to_pupps } from '../../../util/chia';
 import useStandardWallet from '../../../hooks/useStandardWallet';
 import PlotNFTSelectFaucet from './PlotNFTSelectFaucet';
 
@@ -43,10 +43,10 @@ async function prepareSubmitData(data: FormData): SubmitData {
     initialTargetState.relative_lock_height = relative_lock_height;
   }
 
-  const feeMojos = chia_to_mojo(fee);
+  const feePuppss = chia_to_pupps(fee);
 
   return {
-    fee: feeMojos,
+    fee: feePuppss,
     initialTargetState,
   };
 }