+++ chia-blockchain-gui/src/components/trading/TradesTable.tsx
@@ -3,7 +3,7 @@ import { Table } from '@chia/core';
 import styled from 'styled-components';
 import { Trans } from '@lingui/macro';
 import { Box } from '@material-ui/core';
-import { mojo_to_chia_string } from '../../util/chia';
+import { pupps_to_chia_string } from '../../util/chia';
 
 const Amount = styled(Box)`
   white-space: normal;
@@ -46,7 +46,7 @@ export default function TradesTable(props: Props) {
         return {
           side: amount < 0 ? <Trans>Sell</Trans> : <Trans>Buy</Trans>,
           name: <Amount>{name}</Amount>,
-          amount: <Amount>{mojo_to_chia_string(humanAmount)}</Amount>,
+          amount: <Amount>{pupps_to_chia_string(humanAmount)}</Amount>,
         };
       }),
     [rows],