+++ chia-blockchain-gui/src/components/wallet/WalletHistory.tsx
@@ -4,8 +4,8 @@ import { Box, Tooltip, Typography } from '@material-ui/core';
 import { Card, CopyToClipboard, Flex, Table } from '@chia/core';
 import type { Row } from '../core/components/Table/Table';
 import {
-  mojo_to_chia_string,
-  mojo_to_colouredcoin_string,
+  pupps_to_chia_string,
+  pupps_to_colouredcoin_string,
 } from '../../util/chia';
 import { unix_to_short_date } from '../../util/utils';
 import TransactionType from '../../constants/TransactionType';
@@ -57,12 +57,12 @@ const getCols = (type: WalletType) => [
   {
     field: (row: Row) =>
       type === WalletType.COLOURED_COIN
-        ? mojo_to_colouredcoin_string(row.amount)
-        : mojo_to_chia_string(row.amount),
+        ? pupps_to_colouredcoin_string(row.amount)
+        : pupps_to_chia_string(row.amount),
     title: <Trans>Amount</Trans>,
   },
   {
-    field: (row: Row) => mojo_to_chia_string(row.fee_amount),
+    field: (row: Row) => pupps_to_chia_string(row.fee_amount),
     title: <Trans>Fee</Trans>,
   },
 ];