+++ chia-blockchain-gui/src/components/wallet/create/createRLAdmin.jsx
@@ -20,7 +20,7 @@ import {
 } from '../../../modules/createWallet';
 import { useStyles } from './WalletCreate';
 import { create_rl_admin_action } from '../../../modules/message';
-import { chia_to_mojo } from '../../../util/chia';
+import { chia_to_pupps } from '../../../util/chia';
 import { openDialog } from '../../../modules/dialog';
 
 export const customStyles = makeStyles((theme) => ({
@@ -156,12 +156,12 @@ export const CreateRLAdminWallet = () => {
     dispatch(createState(true, true));
     const interval = interval_input.value;
     const interval_value = Number.parseInt(Number(interval));
-    const chiaper = chia_to_mojo(chiaper_input.value);
+    const chiaper = chia_to_pupps(chiaper_input.value);
     const chiaper_value = Number.parseInt(Number(chiaper));
     const userpubkey = userpubkey_input.value;
-    const amount = chia_to_mojo(amount_input.value);
+    const amount = chia_to_pupps(amount_input.value);
     const amount_value = Number.parseInt(Number(amount));
-    // var fee = chia_to_mojo(fee_input.value);
+    // var fee = chia_to_pupps(fee_input.value);
     // TODO(lipa): send fee to server
     // const fee_value = parseInt(Number(fee));
     dispatch(