+++ chia-blockchain-gui/src/components/wallet/create/recoverDIDWallet.js
@@ -24,7 +24,7 @@ import {
 } from '../../../modules/createWallet';
 import ArrowBackIosIcon from '@material-ui/icons/ArrowBackIos';
 import { recover_did_action } from '../../../modules/message';
-import { chia_to_mojo } from '../../../util/chia';
+import { chia_to_pupps } from '../../../util/chia';
 import { openDialog } from '../../../modules/dialog';
 import { useForm, Controller, useFieldArray } from 'react-hook-form';