+++ chia-blockchain-gui/src/components/wallet/did/DIDWallet.js
@@ -31,7 +31,7 @@ import {
 import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
 import { Tooltip } from '@material-ui/core';
 import HelpIcon from '@material-ui/icons/Help';
-import { mojo_to_chia_string } from '../../../util/chia';
+import { pupps_to_chia_string } from '../../../util/chia';
 import { useForm, Controller, useFieldArray } from 'react-hook-form';
 import { openDialog } from '../../../modules/dialog';
 import useCurrencyCode from '../../../hooks/useCurrencyCode';
@@ -474,7 +474,7 @@ const BalanceCardSubSection = (props) => {
           </Box>
           <Box>
             <Typography variant="subtitle1">
-              {mojo_to_chia_string(props.balance)} {currencyCode}
+              {pupps_to_chia_string(props.balance)} {currencyCode}
             </Typography>
           </Box>
         </Box>